+++ tests/public_api/consumer_private_negative/main.cpp
@@ -1,0 +1,6 @@
+#include <QCNetworkReply_p.h>
+
+int main()
+{
+    return 0;
+}